Deal with adding new users to groups, will be changed.
The password is needed when creating new users. Current solution is to
generate the random password for each such instance and thus make sure
that account could not be taken over. A better solution is needed in the
long run. This was a quick fix to get that part of functionality
running.